pyproject.toml
[tool.poetry]
name = "holocal"
version = "0.2.4"
description = "This is an iCalendar generator for Holodule."
authors = ["Sarisia <git@sarisia.cc>", "gemmaro <gemmaro.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ics = "^0.7.2"
python-dotenv = "^0.15.0"
isodate = "^0.7.2"

aiohttp = "^3.11"
aiosqlite = "^0.20" # SQLite backend for aiohttp-client-cache

[tool.poetry.dependencies.aiohttp-client-cache]
version = "^0.12.4"
extras = ["sqlite"]

[tool.poetry.group.dev.dependencies]
sphinx = "^6.2"
myst-parser = "^3.0"
sphinx-intl = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
